Added couple of effects, mainly conditionals. And then made the time bound card use some of them. I admit the new style is really verbose, but I still think it is really neat overall.
File count grew so fast...

It is not *that* big of thing.

I did not test them all but they all should "just work" they are quite simple.
But it will be fun to track down if any bugs arise.

+++ app/src/main/java/com/sq/thed_ck_licker/ecs/components/effectthing/healthEffects/GainMaxHealth.kt
@@ -13,7 +13,7 @@ data class GainMaxHealth(override val amount: Float) : HealthEffect() {
     override fun execute(context: EffectContext): Float {
         val (sourceMulti, targetMulti) = getMultipliers(context)
         val healthComp = (context.target get HealthComponent::class)
-        val amount = (this.amount * sourceMulti * targetMulti).toFloat()
+        val amount = (this.amount * sourceMulti * targetMulti)
         return healthComp.increaseMaxHealth(amount)
     }
 }
